# golang_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Kron-x market",
        "title": "Kron-x App API",
        "contact": {
            "name": "API Creator",
            "url": "https://t.me/amirich18",
            "email": "adamstradvers@gmail.com"
        },
        "version": "2.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/favourite": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get user favourites list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favourite"
                ],
                "summary": "Get user favourites list",
                "operationId": "get_user_favourite",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.DefaultData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.FavouritesProducts"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/favourite/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update user favourites list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favourite"
                ],
                "summary": "Update user favourites list",
                "operationId": "update_user_favourite",
                "parameters": [
                    {
                        "description": "new favourite product info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateFavourite"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/favourite/del": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete item from user favourites list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "favourite"
                ],
                "summary": "Delete item from user favourites list",
                "operationId": "delete_user_favourite",
                "parameters": [
                    {
                        "description": "delete favourite product info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteFavourite"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.DefaultData": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "dto.DeleteFavourite": {
            "type": "object",
            "required": [
                "product_id"
            ],
            "properties": {
                "product_id": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.FavouritesProducts": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "category_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_128": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "vendor_code": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateFavourite": {
            "type": "object",
            "required": [
                "product_id"
            ],
            "properties": {
                "product_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}